pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-music"
version = "0.1.0"
description = "Music retrieval with MERT and CQT baselines"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "torch",
    "torchaudio",
    "librosa",
    "numpy<2.0.0",
    "scipy",
    "scikit-learn",
    "faiss-gpu",
    "transformers",
    "soundfile",
    "pandas",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
